# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Tour Booking API",
    "version": "1.0.0",
    "description": "SRS-TOUR-2026-v1.0. Values in VND; UTC ISO 8601 timestamps."
  },
  "servers": [
    {
      "url": "http://localhost:4000/api/v1"
    }
  ],
  "paths": {
...
    "/tours": {
      "get": {
        "summary": "Chỉ ACTIVE, countryCode VN, chưa soft-delete.",
        "operationId": "get_tours",
        "tags": ["tours"],
        "security": [],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 100
            }
          },
          {
            "name": "destination",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "items": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "title": {
                                "type": "string",
                                "minLength": 3,
                                "maxLength": 150
                              },
                              "slug": {
                                "type": "string",
                                "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$",
                                "maxLength": 150
                              },
                              "description": {
                                "type": "string",
                                "minLength": 10,
                                "maxLength": 10000
                              },
                              "destination": {
                                "type": "string",
                                "minLength": 2,
                                "maxLength": 100
                              },
                              "countryCode": {
                                "type": "string",
                                "enum": ["VN"],
                                "default": "VN"
                              },
                              "durationDays": {
                                "type": "integer",
                                "minimum": 1,
                                "maximum": 60
                              },
                              "status": {
                                "type": "string",
                                "enum": ["DRAFT", "ACTIVE", "INACTIVE"],
                                "default": "DRAFT"
                              },
                              "id": {
                                "type": "string",
                                "format": "uuid"
                              },
                              "createdAt": {
                                "type": "string",
                                "format": "date-time"
                              },
                              "updatedAt": {
                                "type": "string",
                                "format": "date-time"
                              }
                            },
                            "required": [
                              "title",
                              "slug",
                              "description",
                              "destination",
                              "durationDays",
                              "id",
                              "createdAt",
                              "updatedAt"
                            ],
                            "additionalProperties": false
                          }
                        },
                        "page": {
                          "type": "integer"
                        },
                        "pageSize": {
                          "type": "integer"
                        },
                        "total": {
                          "type": "integer"
                        }
                      },
                      "required": ["items", "page", "pageSize", "total"],
                      "additionalProperties": false
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "requestId": {
                          "type": "string"
                        },
                        "timestamp": {
                          "type": "string",
                          "format": "date-time"
                        }
                      },
                      "required": ["requestId", "timestamp"],
                      "additionalProperties": false
                    }
                  },
                  "required": ["data", "meta"],
                  "additionalProperties": false
                },
                "example": {
                  "data": {
                    "items": [
                      {
                        "id": "11111111-1111-4111-8111-111111111111",
                        "title": "Đà Nẵng và Hội An",
                        "slug": "da-nang-hoi-an",
                        "description": "Hành trình khám phá Đà Nẵng và Hội An trong ba ngày.",
                        "destination": "Đà Nẵng",
                        "countryCode": "VN",
                        "durationDays": 3,
                        "status": "ACTIVE",
                        "createdAt": "2026-12-01T01:00:00.000Z",
                        "updatedAt": "2026-12-01T01:00:00.000Z"
                      }
                    ],
                    "page": 1,
                    "pageSize": 20,
                    "total": 1
                  },
                  "meta": {
                    "requestId": "example-request-id",
                    "timestamp": "2026-12-01T01:00:00.000Z"
                  }
                }
              }
            }
          },
          "400": {
            "description": "See error.code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "See error.code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "See error.code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "See error.code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "See error.code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "See error.code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "See error.code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "502": {
            "description": "See error.code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "See error.code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/tours/{id}": {
      "get": {
        "summary": "Chỉ xem được tour công khai.",
        "operationId": "get_tours__id_",
        "tags": ["tours"],
        "security": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "title": {
                          "type": "string",
                          "minLength": 3,
                          "maxLength": 150
                        },
                        "slug": {
                          "type": "string",
                          "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$",
                          "maxLength": 150
                        },
                        "description": {
                          "type": "string",
                          "minLength": 10,
                          "maxLength": 10000
                        },
                        "destination": {
                          "type": "string",
                          "minLength": 2,
                          "maxLength": 100
                        },
                        "countryCode": {
                          "type": "string",
                          "enum": ["VN"],
                          "default": "VN"
                        },
                        "durationDays": {
                          "type": "integer",
                          "minimum": 1,
                          "maximum": 60
                        },
                        "status": {
                          "type": "string",
                          "enum": ["DRAFT", "ACTIVE", "INACTIVE"],
                          "default": "DRAFT"
                        },
                        "id": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "updatedAt": {
                          "type": "string",
                          "format": "date-time"
                        }
                      },
                      "required": [
                        "title",
                        "slug",
                        "description",
                        "destination",
                        "durationDays",
                        "id",
                        "createdAt",
                        "updatedAt"
                      ],
                      "additionalProperties": false
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "requestId": {
                          "type": "string"
                        },
                        "timestamp": {
                          "type": "string",
                          "format": "date-time"
                        }
                      },
                      "required": ["requestId", "timestamp"],
                      "additionalProperties": false
                    }
                  },
                  "required": ["data", "meta"],
                  "additionalProperties": false
                },
                "example": {
                  "data": {
                    "id": "11111111-1111-4111-8111-111111111111",
                    "title": "Đà Nẵng và Hội An",
                    "slug": "da-nang-hoi-an",
                    "description": "Hành trình khám phá Đà Nẵng và Hội An trong ba ngày.",
                    "destination": "Đà Nẵng",
                    "countryCode": "VN",
                    "durationDays": 3,
                    "status": "ACTIVE",
                    "createdAt": "2026-12-01T01:00:00.000Z",
                    "updatedAt": "2026-12-01T01:00:00.000Z"
                  },
                  "meta": {
                    "requestId": "example-request-id",
                    "timestamp": "2026-12-01T01:00:00.000Z"
                  }
                }
              }
            }
          },
          "400": {
            "description": "See error.code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "See error.code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "See error.code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "See error.code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "See error.code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "See error.code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "See error.code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "502": {
            "description": "See error.code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "See error.code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/tours/{id}/schedules": {
      "get": {
        "summary": "Các lịch OPEN trong tương lai.",
        "operationId": "get_tours__id__schedules",
        "tags": ["tours"],
        "security": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "items": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "tourId": {
                                "type": "string",
                                "format": "uuid"
                              },
                              "departureAt": {
                                "type": "string",
                                "format": "date-time"
                              },
                              "totalSeats": {
                                "type": "integer",
                                "minimum": 1,
                                "maximum": 10000
                              },
                              "adultPrice": {
                                "type": "integer",
                                "minimum": 0,
                                "maximum": 99999999
                              },
                              "childPrice": {
                                "type": "integer",
                                "minimum": 0,
                                "maximum": 99999999
                              },
                              "status": {
                                "type": "string",
                                "enum": ["OPEN", "CLOSED"],
                                "default": "OPEN"
                              },
                              "id": {
                                "type": "string",
                                "format": "uuid"
                              },
                              "reservedSeats": {
                                "type": "integer"
                              },
                              "availableSeats": {
                                "type": "integer"
                              },
                              "serverTime": {
                                "type": "string",
                                "format": "date-time"
                              }
                            },
                            "required": [
                              "tourId",
                              "departureAt",
                              "totalSeats",
                              "adultPrice",
                              "childPrice",
                              "id",
                              "reservedSeats",
                              "availableSeats",
                              "serverTime"
                            ],
                            "additionalProperties": false
                          }
                        },
                        "page": {
                          "type": "integer"
                        },
                        "pageSize": {
                          "type": "integer"
                        },
                        "total": {
                          "type": "integer"
                        }
                      },
                      "required": ["items", "page", "pageSize", "total"],
                      "additionalProperties": false
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "requestId": {
                          "type": "string"
                        },
                        "timestamp": {
                          "type": "string",
                          "format": "date-time"
                        }
                      },
                      "required": ["requestId", "timestamp"],
                      "additionalProperties": false
                    }
                  },
                  "required": ["data", "meta"],
                  "additionalProperties": false
                },
                "example": {
                  "data": {
                    "items": [
                      {
                        "id": "22222222-2222-4222-8222-222222222222",
                        "tourId": "11111111-1111-4111-8111-111111111111",
                        "departureAt": "2026-12-15T01:00:00.000Z",
                        "totalSeats": 30,
                        "reservedSeats": 8,
                        "availableSeats": 22,
                        "adultPrice": 3990000,
                        "childPrice": 2490000,
                        "status": "OPEN",
                        "serverTime": "2026-12-01T01:00:00.000Z"
                      }
                    ],
                    "page": 1,
                    "pageSize": 20,
                    "total": 1
                  },
                  "meta": {
                    "requestId": "example-request-id",
                    "timestamp": "2026-12-01T01:00:00.000Z"
                  }
                }
              }
            }
          },
          "400": {
            "description": "See error.code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "See error.code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "See error.code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "See error.code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "See error.code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "See error.code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "See error.code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "502": {
            "description": "See error.code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "See error.code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "refreshCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "refresh_token"
      }
    },
    "schemas": {
      "ApiError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "details": {}
            },
            "required": ["code", "message"],
            "additionalProperties": false
          },
          "meta": {
            "type": "object",
            "properties": {
              "requestId": {
                "type": "string"
              },
              "timestamp": {
                "type": "string",
                "format": "date-time"
              }
            },
            "required": ["requestId", "timestamp"],
            "additionalProperties": false
          }
        },
        "required": ["error", "meta"],
        "additionalProperties": false
      }
    }
  }
}